Bug 1353088: Ensure we don't attempt to process garbage data. r=gerald
MozReview-Commit-ID: AryC34eLjOg

UltraBlame original commit: bf3ed77b00c1d78b23f014776fd76cb3fc11756b

+++ dom/media/mediasource/TrackBuffersManager.h
@@ -234,7 +234,7 @@ private:
   RefPtr<SourceBufferResource> mCurrentInputBuffer;
   RefPtr<MediaDataDemuxer> mInputDemuxer;
   
-  uint32_t mProcessedInput;
+  uint64_t mProcessedInput;
   Maybe<media::TimeUnit> mLastParsedEndTime;
 
   void OnDemuxerInitDone(nsresult);
